Taxonomy: Bypass clearing taxonomy hierarchy for non-hierarchical taxonomies.
Modifies `clean_taxonomy_cache()` to bypass attempting to delete the `{$taxonomy}_children` option for non-hierarchical taxonomies, such as tags.

This removes a needless database query within `delete_option()` to determine if the `{$taxonomy}_children` option is set. It isn't.

Props: chouby, mukesh27, rollybueno, sabernhardt, shailu25, spacedmonkey, westonruter.
Fixes #64090.

Built from https://develop.svn.wordpress.org/trunk@60933


git-svn-id: http://core.svn.wordpress.org/trunk@60269 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-includes/taxonomy.php
@@ -3732,8 +3732,10 @@ function clean_taxonomy_cache( $taxonomy ) {
 	wp_cache_set_terms_last_changed();
 
 	// Regenerate cached hierarchy.
-	delete_option( "{$taxonomy}_children" );
-	_get_term_hierarchy( $taxonomy );
+	if ( is_taxonomy_hierarchical( $taxonomy ) ) {
+		delete_option( "{$taxonomy}_children" );
+		_get_term_hierarchy( $taxonomy );
+	}
 
 	/**
 	 * Fires after a taxonomy's caches have been cleaned.
